+++ src/modules/kiosks/components/KiosksContent.js
@@ -124,7 +124,7 @@ const KiosksContent = ({
     };
 
     if (search || kiosk || kioskStatus) {
-      const name = search ? { name: { $regexI: search } } : {};
+      const name = search ? { name: { $regex: search, $options: 'i' } } : {};
       const kio = kiosk ? { _id: kiosk } : {};
       const door = kioskStatus ? { doorStatus: kioskStatus } : {};