+++ src/modules/kiosks/sagas/exportCsvTempLogs.js
@@ -1,0 +1,46 @@
+import { call, takeLatest } from 'redux-saga/effects';
+import ls from 'lib/localStorage';
+import { TOKEN_STORAGE_KEY } from 'modules/authentication/constants';
+import { exportCsvTempLogs as action } from '../actions';
+
+export function handlerGetProduct(payload) {
+  let link;
+  link = `/api/v1/kiosk/csv/export/temperature/${payload.from}/${payload.to}?kioskId=${payload.kiosk}`;
+  const token = ls.getItem(TOKEN_STORAGE_KEY);
+
+  fetch(link, {
+    method: 'GET',
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+  })
+    .catch(e => {
+      return e;
+    })
+    .then(response => {
+      if (response.status != 200) {
+        alert('Something went wrong. Please try again later!');
+      } else
+        response.blob().then(blob => {
+          let url = window.URL.createObjectURL(blob);
+          let element = document.createElement('a');
+          element.href = url;
+          element.download = `TempLog_${Date.now()}.csv`;
+          document.body.appendChild(element);
+          element.click();
+          document.body.removeChild(element);
+        });
+    });
+}
+
+function* handler({ payload }) {
+  try {
+    yield call(handlerGetProduct, payload);
+  } catch (e) {
+    console.log(e);
+  }
+}
+
+export default function* saga() {
+  yield takeLatest(action, handler);
+}