+++ src/modules/kiosks/sagas/index.js
@@ -19,6 +19,8 @@ import updatePlayList from './updatePlaylist';
 import deletePlayList from './deletePlaylist';
 import exportCsvOrderList from './exportCsvOrders';
 import exportCsvPackList from './exportCsvPacks';
+import exportCsvTempLogs from './exportCsvTempLogs';
+
 export default function* kiosksSaga() {
   yield all([
     fork(getAllKiosks),
@@ -38,5 +40,6 @@ export default function* kiosksSaga() {
     fork(deleteLoadCell),
     fork(exportCsvOrderList),
     fork(exportCsvPackList),
+    fork(exportCsvTempLogs),
   ]);
 }